ORM update

+++ F2/src/main/java/isel/sisinf/grp02/repositories/ICliente_ParticularRepository.java
@@ -4,5 +4,5 @@ import isel.sisinf.grp02.JPAObjects.Cliente_Particular;
 
 import java.util.Collection;
 
-public interface ICliente_ParticularRepository extends IRepository<Cliente_Particular, Collection<Cliente_Particular>, Long>{
+public interface ICliente_ParticularRepository extends IRepository<Cliente_Particular, Collection<Cliente_Particular>, Integer>{
 }
